Create Weapon struct and move enums to it
