+++ CSGOSpectator/.swiftlint.yml
@@ -5,3 +5,4 @@ disabled_rules: # rule identifiers to exclude from running
 # swiftlint rules
 excluded: # paths to ignore during linting. Takes precedence over `included`.
 - Pods
+- CSGOSpectator/Model/Weapon/Weapon.swift
