--- CSGOSpectator/CSGOSpectator/Model/Weapon/WeaponState.swift
@@ -1,14 +1,0 @@
-//
-//  WeaponState.swift
-//  CSGOSpectator
-//
-//  Created by Adam Wienconek on 13.04.2018.
-//  Copyright © 2018 intive. All rights reserved.
-//
-
-import Foundation
-
-enum WeaponState: String, Decodable {
-    case holstered
-    case active
-}
