--- CSGOSpectator/CSGOSpectator/Model/Weapon/WeaponType.swift
@@ -1,21 +1,0 @@
-//
-//  WeaponType.swift
-//  CSGOSpectator
-//
-//  Created by Adam Wienconek on 13.04.2018.
-//  Copyright © 2018 intive. All rights reserved.
-//
-
-import Foundation
-
-enum WeaponType: String, Decodable {
-    case C4
-    case Shotgun
-    case SubmachineGun
-    case MachineGun
-    case Pistol
-    case Knife
-    case Rifle
-    case Grenade
-    case SniperRifle
-}
